# integration-framework/moirai-zeus/src/main/resources/static/swagger/apis/openapi.yaml
openapi: 3.0.0
info:
  title: Moirai Zeus API
  description: This is the API for the Zeus(Ta3) system's access to the subprojects (e.g. Clotho),
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /content/dropdowns/{key}:
    get:
      tags:
        - Content
      summary: retrieves dropdown Content
      description: retrieves dropdown content
      operationId: retrieveDropdownByKey
      parameters:
        - name: key
          in: path
          description: Contents to retrieve by key
          required: true
          schema:
            type: string
        - name: selectOne
          in: query
          description: Provides a Select one option
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Experiment to save
          content:
            'application/json':
              schema:
                items:
                  $ref: '#/components/schemas/dropdownDao'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        404:
          description: Not found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    stringResponse:
      type: object
      properties:
        value:
          type: string
          example: "Success"
...
    dropdownDao:
        type: object
        x-jpaEntity: true
        required:
          - id
          - lookupId
          - value
          - display
        properties:
          id:
            type: string
            example: ""
            x-jpaIdField: true
            x-generatedId: true
          lookupId:
            type: string
            example: "admin.experiment.type"
          value:
            type: string
            example: "dropDown Value"
          display:
            type: string
            example: "Display Value"
            x-jpaColOverride:
              overrideValue: length= 510
          displayOrder:
            type: integer